+++ src/pieces/sortings/index.js
@@ -1,4 +1,13 @@
 import playTime from './play-time';
 import releaseDate from './release-date';
+import globalPlayTime from './global-play-time';
+import trending from './trending';
+import favoriteCount from './favorite-count';
 
-export default { playTime, releaseDate };
+export default {
+  playTime,
+  releaseDate,
+  trending,
+  globalPlayTime,
+  favoriteCount,
+};