+++ src/store/actions/types/update-favorite-count.type.js
@@ -1,0 +1,1 @@
+export default 'UPDATE_FAVORITE_COUNT';