+++ src/store/actions/types/update-global-play-time.type.js
@@ -1,0 +1,1 @@
+export default 'UPDATE_GLOBAL_PLAY_TIME';