+++ src/store/index.js
@@ -14,7 +14,7 @@ import notificationsMiddleware from './middleware/notifications.middleware';
 import STATE_STORAGE_KEY from './middleware/local-storage.middleware/key';
 import installPromptMiddleware from './middleware/install-prompt.middleware';
 import castMiddleware from './middleware/cast.middleware';
-import saveStateBeaconMiddleware from './middleware/save-state-beacon.middleware';
+import fetchStatsMiddleware from './middleware/fetch-stats.middleware';
 import pieces from '../pieces/index';
 import getInitialState from './get-initial-state';
 
@@ -41,8 +41,8 @@ const allMiddlewares = [
   onlineStatusMiddleware,
   notificationsMiddleware,
   installPromptMiddleware,
+  fetchStatsMiddleware,
   localStorageMiddleware,
-  saveStateBeaconMiddleware,
 ];
 const desktopMiddlewares = allMiddlewares.concat([interruptUnloadMiddleware]);