+++ src/store/reducers/is-online.reducer.js
@@ -1,5 +1,5 @@
 import SET_ONLINE_STATUS from '../actions/types/set-online-status.type';
-import getOnlineStatus from '../get-online-status';
+import getOnlineStatus from '@utils/get-online-status';
 
 const isOnlineReducer = (state = getOnlineStatus(), action) =>
   action.type === SET_ONLINE_STATUS ? action.payload : state;