+++ src/store/reducers/root.reducer.js
@@ -21,6 +21,8 @@ import sorting from './sorting.reducer';
 import visiblePieceIdsReducer from './visible-piece-ids.reducer';
 import isInstallable from './is-installable.reducer';
 import cachedPieceIds from './cached-piece-ids.reducer';
+import globalPlayTime from './global-play-time.reducer';
+import favoriteCount from './favorite-count.reducer';
 
 const combinedReducer = combineReducers({
   isMuted,
@@ -44,6 +46,8 @@ const combinedReducer = combineReducers({
   sorting,
   isInstallable,
   cachedPieceIds,
+  globalPlayTime,
+  favoriteCount,
 });
 
 const rootReducer = (state = {}, action) => {