+++ src/utils/map-to-obj.js
@@ -1,0 +1,7 @@
+const mapToObj = map =>
+  Array.from(map).reduce((o, [key, value]) => {
+    o[key] = value;
+    return o;
+  }, {});
+
+export default mapToObj;