+++ src/utils/pipe.js
@@ -1,0 +1,3 @@
+const pipe = (...fns) => x => fns.reduce((y, fn) => fn(y), x);
+
+export default pipe;