+++ docs/CHANGELOG.md
@@ -7,7 +7,19 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-(unreleased)
+(none)
+
+## [1.15.0] - 2019-10-06
+
+### Changed
+
+- Pieces will no longer resort by playtime while a piece is playing
+- Update Patron list
+
+### Added
+
+- New sortings based on global data: trending, global play time, and global favorites
+- New piece: "Soundtrack"
 
 ## [1.14.0] - 2019-09-29
 
@@ -513,7 +525,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Sound files will no longer be fetched and cached during service worker installation. They'll be cached once they are fetched for the first time. This significantly reduces cache usage since only one audio format is used per client.
 
-[unreleased]: https://github.com/generative-music/generative.fm/compare/v1.14.0...HEAD
+[unreleased]: https://github.com/generative-music/generative.fm/compare/v1.15.0...HEAD
+[1.15.0]: https://github.com/generative-music/generative.fm/compare/v1.14.0...v1.15.0
 [1.14.0]: https://github.com/generative-music/generative.fm/compare/v1.13.0...v1.14.0
 [1.13.0]: https://github.com/generative-music/generative.fm/compare/v1.12.0...v1.13.0
 [1.12.0]: https://github.com/generative-music/generative.fm/compare/v1.11.1...v1.12.0